Major milestone: moved IndirectPlayer into vdom
This was the first step to getting multiplayer working with Androminion.  We
need to move the player code into vdom, so someone can start a vdom server on
a machine somewhere that the android client can connect to.  IndirectPlayer
was the largest part of that step.  We still need to move RemotePlayer and
VDomServer over, but I think those will be a little easier.

Note that there are probably still some bugs in this.  I've done some basic
testing, and the app still runs, but there are too many cards for me to test
all them myself without an automated testing framework.

So, the next steps:
1. Move RemotePlayer and VDomServer into the vdom code.
2. Get client to connect to a remote machine, with some way to specify a host
and port (or to start a local server on the android device, as is currently
done).
3. Make sure the JOINGAME events work with multiple players trying to connect.
At this point the game should be able to be played, with two android devices
and a remote host running the server.
4. Figure out a way to specify an android device as the remote host, perhaps
over bluetooth (would it just work over wifi, if you knew the IP address of
the other android device?).  Then all you need is two devices, both with the
app installed, and you can do multiplayer.
5. Implement a UI for hotseat play?